renames actions/types.js to actions/actionTypes.js and updates imports @ reducers && adds suggestions for memory save @ store/configureStore.dev.js && adds ternary expression @ configureStore.js && renames api/huoa.js to apis/content.js
