+++ src/reducers/donationReducer.js
@@ -1,4 +1,4 @@
-//import { CREATE_DONATION, FETCH_DONATIONS, FETCH_DONATION } from '../actions/types';
+//import { CREATE_DONATION, FETCH_DONATIONS, FETCH_DONATION } from '../actions/actionTypes';
 
 export default (state = {}, action) => {
 	switch (action.type) {
