+++ src/reducers/eventReducer.js
@@ -1,4 +1,4 @@
-//import { CREATE_EVENT, FETCH_EVENT, FETCH_EVENTS, EDIT_EVENT, DELETE_EVENT } from '../actions/types';
+//import { CREATE_EVENT, FETCH_EVENT, FETCH_EVENTS, EDIT_EVENT, DELETE_EVENT } from '../actions/actionTypes';
 
 export default (state = {}, action) => {
 	switch (action.type) {
