+++ src/reducers/newsReducer.js
@@ -1,4 +1,4 @@
-//import { CREATE_NEWS, FETCH_NEWS, FETCH_ALL_NEWS, EDIT_NEWS, DELETE_NEWS } from '../actions/types';
+//import { CREATE_NEWS, FETCH_NEWS, FETCH_ALL_NEWS, EDIT_NEWS, DELETE_NEWS } from '../actions/actionTypes';
 
 export default (state = {}, action) => {
 	switch (action.type) {
