+++ src/reducers/subscribeReducer.js
@@ -1,4 +1,4 @@
-//import { CREATE_SUBSCRIPTION, FETCH_SUBSCRIPTIONS, DELETE_SUBSCRIPTION } from '../actions/types';
+//import { CREATE_SUBSCRIPTION, FETCH_SUBSCRIPTIONS, DELETE_SUBSCRIPTION } from '../actions/actionTypes';
 
 export default (state = {}, action) => {
 	switch (action.type) {
