+++ src/store/configureStore.dev.js
@@ -4,8 +4,7 @@ import reducers from '../reducers';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const configureStore = () => {
-	const store = createStore(reducers, composeEnhancers(applyMiddleware(thunk)));
-	return store;
+	return createStore(reducers, composeEnhancers(applyMiddleware(thunk)));
 };
 
 export default configureStore;
