+++ src/store/configureStore.js
@@ -1,5 +1,2 @@
-if (process.env.NODE_ENV === 'production') {
-	module.exports = require('./configureStore.prod');
-} else {
-	module.exports = require('./configureStore.dev');
-}
+module.exports =
+	process.env.NODE_ENV === 'production' ? require('./configureStore.prod') : require('./configureStore.dev');
